SW-4676 Turn the query filters hook into a session filters hook that stores the selected filters into the browser session
